Fxing bugs and improving plugin list pull
* Adding factories required in order to make the new revision metadata
component function.
* Revising the PluginListService to appropriately query the database
for information concerning previously unseen plugins pulled from the
repository.

+++ code/src/Commands/UtilCleanDataCommand.php
@@ -5,11 +5,20 @@ declare(strict_types=1);
 namespace AssetGrabber\Commands;
 
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 class UtilCleanDataCommand extends Command
 {
+    /** @var int[] */
+    private $filesDirsCount = [
+        'files' => 0,
+        'dirs'  => 0,
+        'size'  => 0,
+    ];
+
     protected function configure(): void
     {
         $this->setName('util:clean-data')
@@ -18,6 +27,20 @@ class UtilCleanDataCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $details = $this->countFilesAndDirectories('/opt/assetgrabber/data');
+        $files   = $details['files'];
+        $dirs    = $details['dirs'];
+        $size    = $this->convertBytesToUnit($details['size']);
+
+        /** @var QuestionHelper $question */
+        $question = $this->getHelper('question');
+        $output->writeln("There are $files files and $dirs directories totalling $size.");
+        $q = new ConfirmationQuestion('Are you SURE you want to clean data? THERE WILL BE NO OTHER CONFIRMATION.');
+        if (! $question->ask($input, $output, $q)) {
+            $output->writeln('Aborting!');
+            return self::SUCCESS;
+        }
+
         $output->writeln('Cleaning data directory...');
         $result = $this->deleteFilesAndDirectories('/opt/assetgrabber/data');
         if ($result) {
@@ -52,5 +75,49 @@ class UtilCleanDataCommand extends Command
         }
 
         return true;
+    }
+
+    /**
+     * @return int[]
+     */
+    private function countFilesAndDirectories(string $dir): array
+    {
+        $files = scandir($dir);
+        foreach ($files as $file) {
+            if ($file === '.' || $file === '..' || $file === '.gitkeep') {
+                continue;
+            }
+            if (is_dir($dir . '/' . $file)) {
+                $this->filesDirsCount['dirs']++;
+                $this->countFilesAndDirectories($dir . '/' . $file);
+            } else {
+                $this->filesDirsCount['files']++;
+                $this->filesDirsCount['size'] += filesize($dir . '/' . $file);
+            }
+        }
+
+        return $this->filesDirsCount;
+    }
+
+    private function convertBytesToUnit(int $bytes): string
+    {
+        switch (true) {
+            case $bytes > 1073741824:
+                $size = number_format($bytes / 1073741824, 2) . ' GB';
+                break;
+
+            case $bytes > 1048576:
+                $size = number_format($bytes / 1048576, 2) . ' MB';
+                break;
+
+            case $bytes > 1024:
+                $size = number_format($bytes / 1024, 2) . ' KB';
+                break;
+
+            default:
+                $size = $bytes . ' bytes';
+        }
+
+        return $size;
     }
 }
